Fix PAM not being able to add groups correctly
When trying to use the pam_group module, the groups would not properly
be added when using greetd. It seems like the issue is that "initgroups"
need to be called before "pam_setcred" is called. The man page
explicitly states this. I looked into how gdm does this and made this
mimic the logic used there.